feat: update table

+++ src/pages/dashboard/components/action-list.jsx
@@ -1,19 +1,20 @@
+import {
+    AlertDialog,
+    AlertDialogAction,
+    AlertDialogCancel,
+    AlertDialogContent,
+    AlertDialogDescription,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogTitle,
+    AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+
 import { Button } from "@/components/ui/button";
-import {
-    DropdownMenu,
-    DropdownMenuContent,
-    DropdownMenuGroup,
-    DropdownMenuItem,
-    DropdownMenuLabel,
-    DropdownMenuSeparator,
-    DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-    QueryClient,
-    QueryClientProvider
-} from "@tanstack/react-query";
-import { MoreHorizontal } from "lucide-react";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { FilePenLine, Trash2 } from "lucide-react";
 import PropTypes from "prop-types";
+import { useNavigate } from "react-router-dom";
 
 const queryClient = new QueryClient();
 
@@ -26,30 +27,43 @@ const ActionList = ({ deleteItem, id }) => {
 };
 
 const Action = ({ deleteItem, id }) => {
+    const navigate = useNavigate();
+
+    const handleUpdate = () => {
+        navigate(`/dashboard/table/update/${id}`);
+    };
 
     return (
-        <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-                <Button variant="outline">
-                    <span className="sr-only">Open Menu</span>
-                    <MoreHorizontal className="h-4 w-4" />
-                </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-                <DropdownMenuLabel>Actions</DropdownMenuLabel>
-                <DropdownMenuSeparator />
-                <DropdownMenuGroup>
-                    <DropdownMenuItem>Edit</DropdownMenuItem>
-                    <DropdownMenuItem
-                        onClick={() => {
-                            deleteItem.mutate(id);
-                        }}
-                    >
-                        Delete
-                    </DropdownMenuItem>
-                </DropdownMenuGroup>
-            </DropdownMenuContent>
-        </DropdownMenu>
+        <div className="space-x-4">
+            <Button variant="outline" onClick={handleUpdate}>
+                <FilePenLine className="mr-2 h-4 w-4" /> Edit
+            </Button>
+            <AlertDialog>
+                <AlertDialogTrigger className="bg-zinc-900 text-zinc-50 inline-flex items-center justify-center rounded-md h-10 px-4 py-2">
+                    <Trash2 className="h-4 w-4" />
+                </AlertDialogTrigger>
+                <AlertDialogContent>
+                    <AlertDialogHeader>
+                        <AlertDialogTitle>
+                            Apakah Anda Yakin Ingin Menghapus Data Ini?
+                        </AlertDialogTitle>
+                        <AlertDialogDescription>
+                            Data yang sudah dihapus tidak dapat dikembalikan.
+                        </AlertDialogDescription>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                        <AlertDialogCancel>Cancel</AlertDialogCancel>
+                        <AlertDialogAction
+                            onClick={() => {
+                                deleteItem.mutate(id);
+                            }}
+                        >
+                            Continue
+                        </AlertDialogAction>
+                    </AlertDialogFooter>
+                </AlertDialogContent>
+            </AlertDialog>
+        </div>
     );
 };
